# docs/swagger.yaml
definitions:
...
  go-enterprise-api_internal_domain_role.Role:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  go-enterprise-api_internal_domain_user.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      provider:
        type: string
      provider_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  go-enterprise-api_internal_interface_httpapi.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  go-enterprise-api_internal_interface_httpapi.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorBody'
    type: object
  internal_interface_http_handler.assignPermissionRequest:
    properties:
      permission_id:
        type: integer
    type: object
...
  internal_interface_http_handler.createUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  internal_interface_http_handler.roleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  internal_interface_http_handler.updateUserRequest:
    properties:
      name:
        type: string
      status:
        type: string
    type: object
host: localhost:18080
info:
  contact:
    name: GitOps
  description: 'Enterprise boilerplate: Dynamic Rate Limit, Soft Delete, Audit Logs,
    RBAC, JWT + Redis sessions, Microsoft Entra (OIDC). All requests pass through
    the Nginx gateway (X-Gateway-Token handshake) — public endpoints are also rate-limited
    per IP at the gateway.'
  title: Go Enterprise API
  version: "1.0"
paths:
...
  /api/v1/roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go-enterprise-api_internal_domain_role.Role'
            type: array
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.roleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_role.Role'
        "409":
          description: slug already in use
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - roles
  /api/v1/roles/{id}:
    delete:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete role
      tags:
      - roles
    get:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_role.Role'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get role
      tags:
      - roles
    put:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.roleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_role.Role'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - roles
  /api/v1/roles/{id}/permissions:
    post:
      consumes:
      - application/json
      description: Links a permission to a role. Audited as PERMISSION_GRANTED.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission to grant
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.assignPermissionRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Grant permission to role
      tags:
      - roles
  /api/v1/roles/{id}/permissions/{permissionId}:
    delete:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission ID
        in: path
        name: permissionId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke permission from role
      tags:
      - roles
  /api/v1/roles/{id}/restore:
    post:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore soft-deleted role
      tags:
      - roles
  /api/v1/users:
    get:
      description: Returns all active (non soft-deleted) users. Audited resource;
        RBAC protected.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go-enterprise-api_internal_domain_user.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a local-provider user. The password is bcrypt-hashed in
        the application layer. Will not accept unknown JSON fields.
      parameters:
      - description: User to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "409":
          description: email already in use
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      description: Soft delete (sets deleted_at). The email becomes reusable thanks
        to the partial unique index.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_user.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates name and/or status. Audited (old → new data) inside the
        same transaction.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /api/v1/users/{id}/restore:
    post:
      description: Clears deleted_at. Audited.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore soft-deleted user
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Provide a JWT access token obtained from /api/v1/auth/login or the
      Entra OIDC flow as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"